### 2020-06-07 #2298 @whedon
[PRE REVIEW]: Vizumap: an R package for visualising uncertainty in spatial data
**Submitting author:** @lydialucchesi (<a href="http://orcid.org/0000-0002-1901-4301">Lydia Lucchesi</a>)
**Repository:** <a href="https://github.com/lydialucchesi/Vizumap" target ="_blank">https://github.com/lydialucchesi/Vizumap</a>
**Version:** v1.2.0
