### 2020-06-18 #2298 @arfon
This also might be a good submission for @bstabler to get started with.
