+++ man/build_emap.Rd
@@ -110,7 +110,7 @@ exc_pal <- build_palette(name = "usr", colrange = list(colour = c("yellow", "red
                                                        difC = c(1, 1)))
                                                        view(exc_pal)
 # Create map and view it
-map <- build_emap(data = exc_9596,  shapefile = UB_shp, id = "scID",
+map <- build_emap(data = UB_tss,  shapefile = UB_shp, id = "scID",
             key_label = "Pr[TSS > 837mg/L]")
 view(map)
 
