+++ man/UB_tss.Rd
@@ -1,12 +1,12 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/exc_9596.r
+% Please edit documentation in R/UB_tss.r
 \docType{data}
-\name{exc_9596}
-\alias{exc_9596}
-\title{exc_9596}
+\name{UB_tss}
+\alias{UB_tss}
+\title{UB_tss}
 \format{A data frame comprising 411 rows (one for each RCA) and 5 columns.}
 \usage{
-exc_9596
+UB_tss
 }
 \description{
 A dataset that contains estimates of Total Suspended Sediment (TSS),
