### 2020-06-30 #2298 @danielskatz
@bstabler - FYI, once you've started a review, nothing you do in this pre-review issue will matter, so there's no need to do anything more here.  And I think you know this now, but just in case not, in the pre-review issue:
1 - become the editor: @whedon assign me as editor
2 - find reviewers
3 - assign first reviewer: @whedon assign @reviewer1 as reviewer
4 - assign next reviewer: @whedon add @reviewerer2 as reviewer
5 - repeat 4 if needed
6 - start the review: @whedon start review
 (which closes the pre-review issue)

See https://joss.readthedocs.io/en/latest/editing.html#pre-review for more official text...
