### 2020-06-30 #2298 @sjsrey
sorry but i'm on family vacation for a bit and can't take this one on.
Sergio (Serge) Rey
Founding Director, Center for Geospatial Sciences
Professor, School of Public Policy
University of California, Riverside
sergio.rey@ucr.edu

On Mon, Jun 29, 2020 at 11:12 PM Ben Stabler <notifications@github.com>
wrote:

> and @jsta <https://github.com/jsta>, @sjsrey <https://github.com/sjsrey>,
> @mikejohnson51 <https://github.com/mikejohnson51>, @jayrobwilliams
> <https://github.com/jayrobwilliams>, @mikerspencer
> <https://github.com/mikerspencer>, @GISerDaiShaoqing
> <https://github.com/GISerDaiShaoqing>, would any of you be willing to
> review this submission for JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html.
>
