### 2020-07-01 #2298 @GISerDaiShaoqing
I'm delighted to receive the invitation to review. But I'm a little busy this month, I'm sorry that I may not have more time to review.

------------------
Shaoqing Dai
Ph.D. Candidate
Faculty of Geo-information Science and Earth Observation(ITC), University of Twente, the Netherlands
Hengelosestraat 99, 7500AE  Enschede，the Netherlands
Tel：18850440994
E-mail ：dsq1993qingge@163.com, daishaoqing16@mails.ucas.ac.cn, s.dai@utwente.nl
Homepage: http://gisersqdai.top/mycv/
