### 2020-07-02 #2409 @nuest
Thanks for the chance to contribute to JOSS!

I'll likely give this a go middle of next week.
