### 2020-07-02 #2298 @GISKid
I do not have time this month for a review, but this package looks really
interesting!
