### 2020-07-03 #2409 @lydialucchesi
Hi Ben,

Thanks for the update and finding us several reviewers.

Cheers,
Lydia

On Tue, Jun 30, 2020 at 4:06 PM Ben Stabler <notifications@github.com>
wrote:

> Hi @lydialucchesi <https://github.com/lydialucchesi>, thanks for the
> submittal. I goofed this up since I'm not supposed to assign myself both
> editor and reviewer. I'm working on getting us another editor and/or
> reviewer. Thanks.
>
