### 2020-07-03 #2409 @lydialucchesi
Hi Daniel,

Sounds great - thanks for letting me know.

Best,
Lydia
