### 2020-07-26 #2409 @nuest
## First review notes

### General

- **Authorship**
  - The first author @lydialucchesi has primarily contributed to the JOSS paper - at least judging from the commits. Major code contributions were made by @pkuhnert. The third author has not made any contribution to the software himself, but I assume there is a significant contribution in algorithm design and the idea behind the package. I assume the contributors discussed this, but IMO it would be best to capture this with a short comment by @pkuhnert and Christopher K. Wikle (does not seem to be on GitHub) ? @bstabler what's your take on this?
- **License**
  - The README says "The package Vizumap version 1.1.0 is licensed under GPL-3 (see LICENSE file)", yet the `LICENSE` file contains a "CSIRO Open Source Software License Agreement (GPLv3)", which does not contain the full text of said license, and also references another software (`asremlREplot`) - please clarify in the README, and suggest to include full license text.
  - This license is AFAICT not OSI-approved
  -  I assume there is no way around a non-standard license for you because of funding?
  - The license statement reads strange because it includes the version. Are other versions released under different licenses? The version `1.1.0` also does not match the release version on GitHub (`1.0.0`) nor the version in the `DESCRIPTION` file (`1.2.0`) - please clarify/fix.

### Documentation

- I personally prefer to suggest installation via `remotes`, because it is much quicker to install than `devtools`. Just an idea for the README.
- Both papers referenced in the README are not open access - maybe you can add links to postprints or otherwise deposited copies of the PDFs? I cannot read them.
- The README and papers mention four methods, the vignette mentions only three. Please adjust/clarify.
- The reference in the README uses two different Zenodo DOIs, please fix.
- The commands in the vignette work :+1: There was a slight problem with plotting though on my machine, see sessionInfo output below: The worl map boundaries of Mexico are plottet over Alaska: ![image](https://user-images.githubusercontent.com/1325054/88483177-6ff28500-cf66-11ea-879a-505ec53ef76f.png)
- I get a warning then running the `pixelate` example in the vignette. _Is that expected?_
```r
> pix <- pixelate(ca_geo, id = "region")
Warning message:
In spTransform(shapefile, CRS("+proj=longlat +datum=WGS84")) :
  NULL source CRS comment, falling back to PROJ string
```
- `ca_data$region %in% pix$region` in the vignette would have a shorter output if wrapped with `all(..)`
- Excellent job mentioning which commands might take a bit longer to run in the vignette!
- The vignette assumes readers execute commands from top to bottom, also with the naming of variables. I suggest to use longer variable names for better understandability and readability (not `map` and `map2`)
- **Community guidelines** are missing AFAICS. The `usethis` package can be very helpful to add these.
- **Automated tests** are missing AFAICS. I know tests for visualisations are hard, but you could, and should, have tests for the data preparations functions.

### Paper

- **Statement of need**: could you clarify the targeted users of your package? You mention decision makers as the readers of the created maps, but who creates them? (minor suggestion)
- **State of the field**: please clarify if on other packages (in R) exist.
- "finding methods that add additional elements" - you probably mean more something like "finding methods that can communicate additional elements in an understandable and meaningful way"? You can easily add more elements, but adding userful once seems to be what your methods do.
- "functions can be found inthe package download" - suggest to rephrase to "is available after package installation". It's pretty straightforward these days to get a GitHub pages website for an R package (`pkgdown`, `tic` can help), so if you don't plan to submit the package to CRAN, then having a little website would be great for users. _This is of course unrelated to the JOSS submission_.
- Can you add links to the code that produces the examples in the paper?
- You really have to zoom into the pixel map example to see the pixels. I strongly suggest to show a zoomed in area of the map only, to make the pixel immediately visible for the reader.

### Code

- `DESCRIPTION` file
  - Double field `Maintainer: Maintainer:`
  - `Description` field is very short.
  - `Authors@R` does not match the JOSS paper authors - please clarify.
  - Since the authors are academic, I strongly suggest you [add ORCIDs to the authors list](https://ropensci.org/technotes/2018/10/08/orcid/)! You could also use the contributor roles here to clarify the different contributions (e.g. of Wikle)
- `pixelate`
  - The function paramter and documentation use "shapefile", which is misleading, because it takes `sf` objects as inputs, which could come from any vector file/database format. Suggest to rephrase/rename. This actually applies to the vignette, too. "Map data" would make more sense than "shapefile", right?
  - You do some projection changing in the background -  mention this in the functions docs. _What about users who want to use another projection in their visualisation?_
- I like that you have a lot of developer comments explaining why you do something in the code :+1:
- I could install the package from source. :+1:
- `devtools::check()` completes with one Note (good job!): you should add `paper` directory to `.Rbuildignore`

_I'd be happy to take another look at the source code, if the more general comments are resolved._ :smiley:

### Potentially going beyond JOSS review

- Did you consider integrating with any color-scheme packages in R to help build "good" palettes?
- What does the "uv" in `read.uv` stand for? Maybe a longer function name makes it more memorable/understandable?
- `sp` is the "old way", the current way to work with vector data is the package `sf`. In the paper you write " and we will continue to improve the toolkit to enhance its utility" - it would be great if you could update your functions (or test them, automatic coercion might make a lot of things work already) to play nicely with `sf` objects.
- Is it possible to export an anmination as a GIF? Would be cool to have that in the README or vignette...

### My session info

```r
> sessionInfo()
R version 4.0.2 (2020-06-22)
Platform: x86_64-pc-linux-gnu (64-bit)
Running under: Ubuntu 20.04 LTS

Matrix products: default
BLAS:   /usr/lib/x86_64-linux-gnu/openblas-pthread/libblas.so.3
LAPACK: /usr/lib/x86_64-linux-gnu/openblas-pthread/liblapack.so.3

locale:
 [1] LC_CTYPE=en_US.UTF-8       LC_NUMERIC=C               LC_TIME=de_DE.UTF-8        LC_COLLATE=en_US.UTF-8
 [5] LC_MONETARY=de_DE.UTF-8    LC_MESSAGES=en_US.UTF-8    LC_PAPER=de_DE.UTF-8       LC_NAME=C
 [9] LC_ADDRESS=C               LC_TELEPHONE=C             LC_MEASUREMENT=de_DE.UTF-8 LC_IDENTIFICATION=C

attached base packages:

other attached packages:
[1] Vizumap_1.2.0

 [1] spbabel_0.5.1        httr_1.4.2           maps_3.3.0           pkgload_1.1.0        tidyr_1.1.0
 [6] bit64_0.9-7          jsonlite_1.7.0       assertthat_0.2.1     sp_1.4-2             animation_2.6
[11] blob_1.2.1           remotes_2.1.1        sessioninfo_1.1.1    pillar_1.4.6         RSQLite_2.2.0
[16] backports_1.1.8      lattice_0.20-41      glue_1.4.1           digest_0.6.25        colorspace_1.4-1
[21] plyr_1.8.6           clisymbols_1.2.0     pkgconfig_2.0.3      devtools_2.3.0       broom_0.7.0
[26] purrr_0.3.4          scales_1.1.1         processx_3.4.3       jpeg_0.1-8.1         ggmap_3.0.0
[31] tibble_3.0.3         farver_2.0.3         generics_0.0.2       cranlike_1.0.2       ggplot2_3.3.2
[36] usethis_1.6.1        ellipsis_0.3.1       withr_2.2.0          cli_2.0.2            magrittr_1.5
[41] crayon_1.3.4         memoise_1.1.0        maptools_1.0-1       ps_1.3.3             fs_1.4.2
[46] fansi_0.4.1          crancache_0.0.0.9001 parsedate_1.2.0      foreign_0.8-79       pkgbuild_1.1.0
[51] tools_4.0.2          prettyunits_1.1.1    RgoogleMaps_1.4.5.3  lifecycle_0.2.0      stringr_1.4.0
[56] munsell_0.5.0        geoaxe_0.1.0         callr_3.4.3          compiler_4.0.2       rlang_0.4.7
[61] debugme_1.1.0        grid_4.0.2           rstudioapi_0.11      rjson_0.2.20         rappdirs_0.3.1
[66] prompt_1.0.0         labeling_0.3         bitops_1.0-6         testthat_2.3.2       gtable_0.3.0
[71] dadjoke_0.1.2        DBI_1.1.0            curl_4.3             reshape2_1.4.4       rematch2_2.1.2
[76] R6_2.4.1             gridExtra_2.3        rgdal_1.5-12         dplyr_1.0.0          rgeos_0.5-3
[81] bit_1.1-15.2         rprojroot_1.3-2      desc_1.2.0           stringi_1.4.6        Rcpp_1.0.5
[86] vctrs_0.3.2          png_0.1-7            tidyselect_1.1.0
```
