### 2020-07-29 #2409 @lydialucchesi
Hi Dr. Nüst (@nuest),

Thank you for reviewing the Vizumap paper and package. I will start
working on the revisions. Is there a date that I should have the
revisions completed by (@bstabler)?

Following is clarification regarding author contributions and licensing.

Chris and I collaborated on the development of three Vizumap
visualisation methods. Chris helped design the first three map types
and provided advice on method implementation (e.g., using the rotation
matrix to implement the glyph map). These maps are described in the
first paper referenced, which can be accessed here:
http://faculty.missouri.edu/~wiklec/LucchesiWikle2017Stat.
Accompanying that paper is a supplementary R vignette that includes
the code I wrote to produce the paper figures:
http://faculty.missouri.edu/~wiklec/sta4150-sup-0001-supplementary.

Petra and I then started a collaboration in 2017 where we translated
the methods and code from the Stat paper into an R package, and this
was housed in a CSIRO Bitbucket repository. Petra made contributions
towards speed up and design of the code to make it more user friendly
and efficient. Petra also contributed the 4th visualisation method in
the package (the exceedance probability map), which was outlined in
the second paper we reference. We are working on creating an
open-access link for this paper.

When we decided to port this code to GitHub and make it public, we had
trouble getting a CSIRO Bitbucket repository transferred to GitHub. We
therefore created a new GitHub repository that unfortunately does not
contain all of the history of commits. If it would be helpful, we can
make the original Bitbucket repository available to you as a separate
repository so you can see the version history and history of commits
to validate contributions.

The License file should not appear in that repository, and the
Description file should just include a reference to GPLv3 only. My
apologies for the error and causing confusion. I will push a change to
the repo to fix this.

Thank you for providing helpful feedback that will improve the paper
and the package.

Best regards,
Lydia
