### 2020-07-29 #2409 @nuest
Just Daniel, please - no "Dr." yet :man_shrugging:

**Thanks for the clarification!** All makes sense - for me a short comment about the project history, e.g. in the contributors documentation, or the README, would suffice to make this clear to future users/readers.
