### 2020-08-07 #2409 @bstabler
@nuest wrote "The first author @lydialucchesi has primarily contributed to the JOSS paper - at least judging from the commits. Major code contributions were made by @pkuhnert. The third author has not made any contribution to the software himself, but I assume there is a significant contribution in algorithm design and the idea behind the package. I assume the contributors discussed this, but IMO it would be best to capture this with a short comment by @pkuhnert and Christopher K. Wikle (does not seem to be on GitHub) ? @bstabler what's your take on this?"

I agree, it would be nice to document the contribution in some traceable way.
