### 2020-08-30 838695e
description file revisions, readme updates, deletion of wrong license file
