+++ README.Rmd
@@ -40,25 +40,23 @@ Petra Kuhnert, CSIRO Data61, Email: Petra.Kuhnert@data61.csiro.au
 
 ## About the Package
 
-Approaches for visualizing uncertainty in spatial data are presented in this package. These include the three approaches developed in Lucchesi and Wikle (2017) and a fourth approach presented in Kuhnert et al. (2018) that uses exceedance probabilities to convey uncertainty on a map (or stream network) from a spatio-temporal model.
+Approaches for visualizing uncertainty in spatial data are presented in this package. These include the three approaches developed in [Lucchesi and Wikle (2017)](http://faculty.missouri.edu/~wiklec/LucchesiWikle2017Stat) and a fourth approach presented in [Kuhnert et al. (2018)](https://publications.csiro.au/publications/#publication/PIcsiro:EP168206).
 
 ### Bivariate Maps
 
-This approach is an extension of previous implementations of choropleth mapping that includes new features to make the maps more useful for interpretation. Bivariate choropleth maps explore the “blending” of two colour schemes, one representing the estimate and a second representing the margin of error, which can be conveyed on a map through a single blended colour. 
-Bivariate colour maps can be created for both areal and point level data. Using a bivariate colour grid, estimates from a spatial
-model (e.g. mean) and the uncertainty surrounding these values (e.g. standard deviation) are mapped simulataneously. Vizumap allows the user to create a bivariate colour scheme by mathematically blending two single hue colour palettes and organising them on a 3 x 3 grid. Users can develop their own bivariate colour palette as well as select from some pre-prepared palettes.
-
+In these bivariate choropleth maps, two colour schemes, one representing the estimates and another representing the margins of error, are blended so that an estimate and its error can be conveyed on a map using a single colour. 
 
 ### Map Pixelation
 
-This approach uses map pixelation to convey uncertainty. This is a novel approach that pixelates areas within each region of the map according to the size of the margin of error. Regions that appear visually as a solid piece of colour reflect smaller margins of error, while more pixelated regions indicate larger margins of error. These maps can be animated using visuanimation techniques to provide a novel “user experience” to visualizing uncertainty on a map. 
+In this novel approach, each map region is pixelated, and then each pixel is filled with a colour from the gradient that falls within the estimate's margin of error. Regions that appear visually as a solid colour reflect smaller margins of error, while more pixelated regions indicate greater uncertainty. These maps can be animated using visuanimation techniques to provide a novel uncertainty visualisation experience.
 
 ### Glyph Rotation
 
-Glyph rotation uses a glyph to represent uncertainty. When rotated in different ways, the rotation of the glyph corresponds to the margin of error. 
+In this method, glyphs located at region centroids are rotated to represent uncertainty. The colour filling each glyph corresponds to the estimate.
 
 ### Exceedance Probability Maps
-The final map based exploration of uncertainty is through exceedance probabilities, that can be showcased on a map to highlight regions that exhibit varying levels of departure from a threshold of concern or target. 
+
+The final map-based exploration is through exceedance probabilities, which are visualised on a map to highlight regions that exhibit varying levels of departure from a threshold of concern or target. 
 
 ## Examples
 
@@ -70,11 +68,11 @@ vignette("Vizumap")
 
 ## How to Reference
 
-Vizumap: An R package for visualizing uncertainty in spatial data, DOI: 10.5281/zenodo.1013930, https://zenodo.org/record/1479951#.Xnr_Qy1L2L8
[…]
