+++ DESCRIPTION
@@ -1,14 +1,16 @@
 Package: Vizumap
 Type: Package
-Title: Visualizing Uncertainty in spatial data
-Version: 1.2.0
+Title: Visualizing uncertainty in spatial data
+Version: 1.1.0
 Author: Lydia Lucchesi and Petra Kuhnert
-Maintainer: Maintainer: Lydia <Lydia.Lucchesi@anu.edu.au>
-Description:
-    Vizumap provides visualization approaches for spatial data.
+Maintainer: Lydia <Lydia.Lucchesi@anu.edu.au>
+Description: Vizumap provides four uncertainty visualization approaches for spatial data.
+    These approaches include the bivariate choropleth map, map pixelation, glyph rotation, and the exceedance probability map.
+    In Vizumap, there are three different types of functions - formatting, building, and viewing functions - that make it easy to create 
+    these uncertainty maps for a range of applied spatial problems.
 Encoding: UTF-8
 Date: 2020-06-04
-Authors@R: c(person("Lydia", "Lucchesi", email = "Lydia.Lucchesi@anu.edu.au", role = "aut"), person("Petra", "Kuhnert", email = "Petra.Kuhnert@data61.csiro.au", role = c("aut", "cre")))
+Authors@R: c(person("Lydia", "Lucchesi", email = "Lydia.Lucchesi@anu.edu.au", role = c("aut", "cre"), comment = "Worked on the development/design of three Vizumap visualisation methods (bivariate map, map pixelation, and glyph rotation) with Chris. Wrote original R code to produce the first example maps for those three methods (http://faculty.missouri.edu/~wiklec/sta4150-sup-0001-supplementary). Developed Vizumap package with Petra."), person("Petra", "Kuhnert", email = "Petra.Kuhnert@data61.csiro.au", role = c("aut"), comment = "Developed Vizumap package with Lydia. Designed the exceendance probability mapping method."), person("Christopher", "Wikle", email = "wiklec@missouri.edu", role = c("ctb"), comment = "Worked on the development/design of three Vizumap visualisation methods (bivariate map, map pixelation, and glyph rotation) with Lydia."))
 Depends:
     R (>= 3.5.0)
 Imports:
@@ -28,7 +30,7 @@ Imports:
     sp,
     spbabel,
     maps
-License: GPL-3 + file LICENSE
+License: GPL-3
 LazyLoad: TRUE
 LazyData: TRUE
 RoxygenNote: 7.0.0
