+++ README.md
@@ -7,11 +7,10 @@
 
 ## Installation
 
-You can install a development version of the Vizumap package from
-[GitHub](https://github.com/lydialucchesi/Vizumap)
+You can install a development version of the Vizumap package using the
+command below.
 
-    # install.packages("devtools")
-    devtools::install_github(repo = "lydialucchesi/Vizumap", build_vignettes = TRUE, force = TRUE)
+    remotes::install_github(repo = "lydialucchesi/Vizumap", build_vignettes = TRUE, force = TRUE)
 
 ## Authors
 
@@ -38,13 +37,12 @@ single colour.
 
 ### Map Pixelation
 
-In this novel approach, each map region is pixelated, and then each
-pixel is filled with a colour from the gradient that falls within the
-estimate’s margin of error. Regions that appear visually as a solid
-colour reflect smaller margins of error, while more pixelated regions
-indicate greater uncertainty. These maps can be animated using
-visuanimation techniques to provide a novel uncertainty visualisation
-experience.
+In this novel approach, each map region is pixelated. Pixels are filled
+with colours representing values within an estimate’s margin of error.
+Regions that appear as a solid colour reflect smaller margins of error,
+while more pixelated regions indicate greater uncertainty. These maps
+can be animated using visuanimation techniques to provide a novel
+uncertainty visualisation experience.
 
 ### Glyph Rotation
 
@@ -67,8 +65,7 @@ relating to each of the visualization methods.
 
 ## How to Reference
 
-Vizumap: An R package for visualizing uncertainty in spatial data, DOI:
-10.5281/zenodo.1013930,
+Vizumap: An R package for visualizing uncertainty in spatial data,
 <https://zenodo.org/record/1479951#.Xnr_Qy1L2L8>.
 
 ## License
@@ -77,12 +74,14 @@ Vizumap version 1.1.0 is licensed under GPLv3.
 
 ## References
 
-<strong>Kuhnert, P.M.</strong>, Pagendam, D.E., Bartley, R., Gladish,
-D.W., Lewis, S.E. and Bainbridge, Z.T. (2018) Making management
-decisions in face of uncertainty: a case study using the Burdekin
-catchment in the Great Barrier Reef, Marine and Freshwater Research, 69,
-1187-1200, <https://doi.org/10.1071/MF17237>.
+Kuhnert, P.M., Pagendam, D.E., Bartley, R., Gladish, D.W., Lewis, S.E.
+and Bainbridge, Z.T. (2018) [Making management decisions in face of
+uncertainty: a case study using the Burdekin catchment in the Great
+Barrier Reef, Marine and Freshwater
+Research](https://publications.csiro.au/publications/#publication/PIcsiro:EP168206),
+69, 1187-1200, <https://doi.org/10.1071/MF17237>.
 
-<strong>Lucchesi, L.R.</strong> and Wikle C.K. (2017) Visualizing
-uncertainty in areal data with bivariate choropleth maps, map pixelation
-and glyph rotation, Stat, <https://doi.org/10.1002/sta4.150>.
+Lucchesi, L.R. and Wikle C.K. (2017) [Visualizing uncertainty in areal
+data with bivariate choropleth maps, map pixelation and glyph
+rotation](http://faculty.missouri.edu/~wiklec/LucchesiWikle2017Stat),
+Stat, <https://doi.org/10.1002/sta4.150>.
