+++ README.Rmd
@@ -39,35 +39,31 @@ Petra Kuhnert, CSIRO Data61, Email: Petra.Kuhnert@data61.csiro.au
 
 ## About the Package
 
-Approaches for visualizing uncertainty in spatial data are presented in this package. These include the three approaches developed in [Lucchesi and Wikle (2017)](http://faculty.missouri.edu/~wiklec/LucchesiWikle2017Stat) and a fourth approach presented in [Kuhnert et al. (2018)](https://publications.csiro.au/publications/#publication/PIcsiro:EP168206).
+Approaches for visualising uncertainty in spatial data are presented in this package. These include the three approaches developed in [Lucchesi and Wikle (2017)](http://faculty.missouri.edu/~wiklec/LucchesiWikle2017Stat) and a fourth approach presented in [Kuhnert et al. (2018)](https://publications.csiro.au/publications/#publication/PIcsiro:EP168206).
 
-### Bivariate Maps
+#### Bivariate Maps
 
 In these bivariate choropleth maps, two colour schemes, one representing the estimates and another representing the margins of error, are blended so that an estimate and its error can be conveyed on a map using a single colour. 
 
-### Map Pixelation
+#### Map Pixelation
 
-In this novel approach, each map region is pixelated. Pixels are filled with colours representing values within an estimate's margin of error. Regions that appear as a solid colour reflect smaller margins of error, while more pixelated regions indicate greater uncertainty. These maps can be animated using visuanimation techniques to provide a novel uncertainty visualisation experience.
+In this approach, each map region is pixelated. Pixels are filled with colours representing values within an estimate's margin of error. Regions that appear as a solid colour reflect smaller margins of error, while more pixelated regions indicate greater uncertainty. These maps can be animated to provide a novel uncertainty visualisation experience.
 
-### Glyph Rotation
+#### Glyph Rotation
 
 In this method, glyphs located at region centroids are rotated to represent uncertainty. The colour filling each glyph corresponds to the estimate.
 
-### Exceedance Probability Maps
+#### Exceedance Probability Maps
 
 The final map-based exploration is through exceedance probabilities, which are visualised on a map to highlight regions that exhibit varying levels of departure from a threshold of concern or target. 
 
 ## Examples
 
-A vignette for the Vizumap package is available and contains examples relating to each of the visualization methods.
+A vignette for the Vizumap package is available and contains examples relating to each of the visualisation methods.
 
 ```
 vignette("Vizumap")
 ```
-
-## How to Reference
-
-Vizumap: An R package for visualizing uncertainty in spatial data, https://zenodo.org/record/1479951#.Xnr_Qy1L2L8.
 
 ## License
 
