+++ vignettes/Vizumap.Rmd
@@ -1,6 +1,6 @@
 ---
 title: "Visualising Uncertainty in Poverty Rates across the USA"
-author: "Petra Kuhnert and Lydia Lucchesi"
+author: "Lydia Lucchesi and Petra Kuhnert"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
@@ -16,11 +16,11 @@ library(ggplot2)
 library(dplyr)
 ```
 
-**Vizumap** is a tool for visualising uncertainty in spatial data. This vignette demonstrates how to use **Vizumap** to create bivariate maps, pixel maps and glyph maps, the three uncertainty visualisation methods developed in Lucchesi and Wikle (2017). We have tried to generalise the approaches so they can be applied to most types of spatial data.
+**Vizumap** is a tool for visualising uncertainty in spatial data. This vignette demonstrates how to use the package to create bivariate maps, pixel maps, glyph maps, and exceedance probability maps. We have tried to generalise the visualisation approaches so that they are applicable for most types of spatial data.
 
-There are several types of functions in **Vizumap**: formatting functions, building functions and viewing functions. Formatting functions prepare data frames and spatial polygons for use in building functions. Building functions create maps, animations, keys and colour palettes. Viewing functions plot these objects. Some of these functions are useful across all three map types, such as `read.uv`, and others are map specific, such as `pixelate`. Below is a table displaying the functions that can be used with each map type. Blue functions format. Green functions create. Red functions plot.
+There are three types of functions in **Vizumap**: formatting functions, building functions, and viewing functions. Formatting functions prepare data frames and spatial polygons for use in building functions. Building functions create maps, animations, keys, and colour palettes. Viewing functions plot these objects. Some of these functions are useful across all four map types, such as `read.uv`, and others are map specific, such as `pixelate`. Below is a table displaying the functions that can be used with each map type. Blue functions format. Green functions create. Red functions plot.
 
-```{r echo = FALSE, fig.align='center', fig.height=4, fig.width=6, fig.asp=.6, warning=FALSE}
+```{r echo = FALSE, fig.align='center', fig.height=5, fig.width=8, fig.asp=.6, warning=FALSE}
 
 b <- data.frame(name = rev(c("read.uv", "build_bmap", "build_bkey", "build_palette", "view", "attach_key")),
                 x = rep(0, 6), 
@@ -28,19 +28,24 @@ b <- data.frame(name = rev(c("read.uv", "build_bmap", "build_bkey", "build_palet
                 col = rev(c("midnightblue", "darkgreen", "darkgreen", "darkgreen", "darkred", "darkred")))
 
 p <- data.frame(name = rev(c("read.uv", "pixelate", "build_pmap", "animate", "view", "")), 
-                x = rep(2, 6), 
+                x = rep(5, 6), 
                 y = seq(0, 5, 1), 
                 col = rev(c("midnightblue", "midnightblue", "darkgreen", "darkgreen", "darkred", "darkred")))
 
 
 g <- data.frame(name = rev(c("read.uv", "build_gmap", "build_gkey", "view", "attach_key", "")), 
-                x = rep(4, 6), 
+                x = rep(10, 6), 
                 y = seq(0, 5, 1),
                 col = rev(c("midnightblue", "darkgreen", "darkgreen", "darkred", "darkred", "darkred")))
 
-functions <- rbind(b, p, g)
+e <- data.frame(name = rev(c("read.uv", "build_emap", "view", "", "", "")), 
[…]
