### 2020-08-30 52ba1c0
ignored paper dir, added paper figure code, switched from commenting to dontrun
