+++ R/attach_key.R
@@ -10,12 +10,11 @@
 #'data(us_geo)
 #'poverty <- read.uv(data = us_data, estimate = "pov_rate", error = "pov_moe")
 #'
-#'#bivariate map and key together
+#'# bivariate map and key together
 #'map <- build_bmap(data = poverty, shapefile = us_geo, id = "GEO_ID",
 #'  border = "state", terciles = TRUE)
 #'key <- build_bkey(data = poverty, terciles = TRUE)
 #'attach_key(map, key)
-#'
 #' @export
 #' @importFrom "gridExtra" "grid.arrange"
 
