+++ paper/paper.md
@@ -27,8 +27,7 @@ bibliography: paper.bib
 # Summary
 
 To make a sound data-driven decision, it is necessary to consider the quality and strength of the evidence. Therefore, it is important that the uncertainty of statistical estimates be effectively communicated to decision makers to ensure it is properly considered in the decision-making 
-process. Generally, this uncertainty information is shared through visualisation (e.g., error bars). However, in spatial applications, finding methods that add additional elements to a map of spatial estimates can be challenging. To address this visualisation shortcoming in spatial 
-statistics, we developed the `Vizumap` R package, which is a toolkit for visualising uncertainty in spatial data.
+process. Generally, this uncertainty information is shared through visualisation (e.g., error bars). However, in spatial applications, finding methods that can communicate additional information, about the spatial estimates, in an understandable and meaningful way can be challenging. To address this visualisation shortcoming in spatial statistics, we developed the `Vizumap` R package. It is a toolkit designed for statisticians, scientists, data journalists, etc., discussing uncertainty in spatial data.
 
 `Vizumap` contains a series of straightforward functions that can be used to create four different types of maps, which are discussed in detail in @vizMethod and @gbrData. The visualisation methods include the bivariate choropleth map, pixel map, glyph map, and exceedance probability map. Following is a brief description of each method.
 
@@ -40,7 +39,7 @@ statistics, we developed the `Vizumap` R package, which is a toolkit for visuali
 
 * Exceedance probability maps show the probability of exceeding a nominated threshold of concern. These probabilities can be pre-calculated and passed to the exceedance probability map function or calculated within the function given a prescribed probability distribution function, estimate, and error. For example, let $X$ denote a random variable, and $x$ is a possible value of that random variable, $X$. $F_x(x)$ is the cumulative distribution function and equals the probability that the value of $X$ is less than or equal to a specific value or threshold, $x$, so $$F_x(x) = Pr[X {\leq} x].$$ The probability that the value $X$ is greater than the specific threshold $x$ can then be written as $$Pr[X > x] = 1 - F_x(x) = 1 - Pr[X {\leq} x].$$ In R, assuming an exponential distribution, this can be evaluated through the following expressions: `1 - pexp(q, rate)` or `pexp(q, rate, lower.tail = FALSE)`.
 
-A comprehensive vignette that demonstrates how to use the `Vizumap` functions can be found in the package download. The functions are divided into three categories: formatting, building, and viewing. Formatting functions prepare data for use in the building functions, which are used to build the colour palettes, maps, and map keys. Viewing functions are used to check and combine the different graphical objects designed with the building functions. Previous applications of the `Vizumap` R package include visualising American Community Survey estimates with their corresponding margins of error. We believe `Vizumap` is useful in a wide range of applications, and we will continue to improve the toolkit to enhance its utility. 
+A comprehensive vignette that demonstrates how to use `Vizumap` is available after package installation. The functions are divided into three categories: formatting, building, and viewing. Formatting functions prepare data for use in the building functions, which are used to build the colour palettes, maps, and map keys. Viewing functions are used to check and combine the different graphical objects designed with the building functions. Previous applications of the `Vizumap` R package include visualising American Community Survey estimates with their corresponding margins of error. We believe `Vizumap` is useful in a wide range of applications, and we will continue to improve the toolkit to enhance its utility. 
 
[…]
