### 2020-09-02 #2409 @lydialucchesi
I am working on revisions and have a running list of what has been done so far saved in the paper folder in the Vizumap repo: https://github.com/lydialucchesi/Vizumap/blob/master/paper/revisionList.docx.
