### 2020-09-14 7867c1e
changed shapefile argument, reformatted comments
