### 2020-09-16 #2409 @nuest
@lydialucchesi Thanks for the update!

I strongly recommend not to put docx files under version control - it is effectively a ZIP archive that git cannot effectively track changes over time with. Why not use a plain text file, with Markdown formatting?
