### 2020-09-16 #2409 @lydialucchesi
@nuest Good idea. I have made the switch, and the new file can be found here: https://github.com/lydialucchesi/Vizumap/blob/master/paper/revisions.md. Thanks, Lydia
