### 2020-09-17 8d38263
added license, code of conduct, and contributing guidelines
