+++ CONTRIBUTING.md
@@ -1,0 +1,33 @@
+# Contributing
+
+Thank you for your cooperation and contribution! We appreciate your help making Vizumap the best it can be.
+
+## How to Propose a Change to Vizumap
+
+### Prerequisites
+
+Before you make a substantial pull request, you should always file an issue and
+make sure someone from the team agrees that it is a problem. If you have found a
+bug, create an associated issue and illustrate the bug with a minimal 
+[reprex](https://www.tidyverse.org/help/#reprex).
+
+### Pull request process
+
+Guidelines for contributors:
+
+* Create a named branch (not master) which merges cleanly with master.
+* Limit the scope of the pull request to a clearly defined improvement or fix.
+* Follow the tidyverse [style guide](https://style.tidyverse.org). Please do not restyle code unrelated to your PR.
+* Use [roxygen2](https://cran.r-project.org/package=roxygen2) for documentation.
+* Use [testthat](https://cran.r-project.org/package=testthat). Contributions with test cases included are easier to accept.
+* For user-facing changes, add a bullet to the top of `NEWS.md` below the
+current development version header describing the changes you made followed by your
+GitHub username and links to relevant issue(s)/PR(s).
+* After you merge a pull request, delete the associated branch if you can. You will see an option for this on the GitHub pull request page.
+
+Every pull request will be reviewed by @lydialucchesi or @pkuhnert. They will either merge it or provide feedback on issues to be addressed. They may add additional commits to the pull request. @lydialucchesi and @pkuhnert can make minor edits to a pull request (e.g., fixing typos) before merging it.
+
+### Code of Conduct
+
+Please note that the Vizumap project is released with a [Contributor Code of Conduct](CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+
