+++ paper/revisions.md
@@ -6,9 +6,13 @@
 * Fixed license description
 * Switched from devtools to remotes installation
 * Removed Zenodo reference
+* Added section about contributing to the package
+  * Generated a Code of Conduct file with usethis::use_code_of_conduct()
+  * Added contribution guidelines file
 
 ### License
 * Removed incorrect file from repo
+* Added correct full text license file (GPLv3)
 
 ### Description
 * Corrected package version (it should be v1.1.0)
@@ -39,6 +43,7 @@
   * “functions can be found in the package download” to “is available after package installation”
 * Clarified audience for statement of need requirement:
   * “However, in spatial applications, finding methods that can communicate additional information, about the spatial estimates, in an understandable and meaningful way can be challenging. To address this visualisation shortcoming in spatial statistics, we developed the Vizumap R package. **It is a toolkit designed for statisticians, scientists, data journalists, etc., discussing uncertainty in spatial data.**”
+
 
 
 
