### 2020-10-09 41dcde0
updated projection part in pixelate
