### 2020-10-09 a2e9801
tests for data formatting
