### 2020-10-11 22b5c2f
updated pixel map figure
