### 2020-10-11 b6f98d7
pixel map figure and text updates, code formatting
