+++ README.Rmd
@@ -67,9 +67,9 @@ vignette("Vizumap")
 
 ## Contribute
 
-Please note that the `Vizumap` project is released with a [Contributor Code of Conduct](CONDUCT.md). By contributing to this project, you agree to abide by its terms.
-
 To contribute to `Vizumap`, please follow these [guidelines](CONTRIBUTING.md).
+
+Please note that the `Vizumap` project is released with a [Contributor Code of Conduct](CONDUCT.md). By contributing to this project, you agree to abide by its terms.
 
 ## License
 
