+++ paper/paper.bib
@@ -34,3 +34,19 @@ pages = {1590-1618},
 doi = {10.1214/16-AOAS950},
 publisher={Institute of Mathematical Statistics}
 }
+
+@Manual{biscalePackage,
+title = {biscale: Tools and Palettes for Bivariate Thematic Mapping},
+author = {Christopher Prener and Timo Grossenbacher and Angelo Zehr},
+year = {2020},
+note = {R package version 0.2.0},
+url = {https://CRAN.R-project.org/package=biscale},
+}
+
+@Manual{pixelatePackage,
+title = {pixelate: Pixelate spatial predictions as per their average uncertainty},
+author={Aimee R Taylor and James A Watson and Caroline O Buckee},
+year = {2020},
+note = {R package version 0.0.0.9000},
+url = {https://github.com/artaylor85/pixelate},
+}
