+++ paper/paper.md
@@ -72,6 +72,6 @@ The exceedance probability map plots the calculated probability of exceeding a s
 
 We would like to acknowledge the support and funding from the CSIRO Digiscape Future Science Platform.
 
-`Vizumap` was built using the following R packages: `ggplot2`, `animation`, `broom`, `dplyr`, `geoaxe`, `ggmap`, `grDevices`, `gridExtra`, `maps`, `maptools`, `plyr`, `reshape2`, `rgdal`, `rgeos`, `roxygen2`, `sp`, `spbabel`,`testthat`, `usethis`, and `utils`.
+`Vizumap` was built using the following R packages: `ggplot2`, `animation`, `broom`, `dplyr`, `geoaxe`, `ggmap`, `grDevices`, `gridExtra`, `maps`, `maptools`, `plyr`, `reshape2`, `rgdal`, `rgeos`, `roxygen2`, `sp`, `spbabel`, `testthat`, `usethis`, and `utils`.
 
 # References
