### 2020-10-14 #2409 @lydialucchesi
Hi @bstabler and @nuest, we have addressed and incorporated all of revisions from the first review (https://github.com/lydialucchesi/Vizumap/blob/master/paper/revisions.md). Please let me know if there is any additional information I can provide regarding the changes made.

Should we update the versioning from 1.0.0 to 2.0.0 given the major changes? Version 1.0.0 is from the original package release in 2017, and the name of the package at that time was VizU.

@nuest, thank you for your thoughtful, constructive feedback in the review. We are so excited with the improved state of the package and paper.

Kind regards, Lydia
