### 2020-10-22 #2409 @nuest
@lydialucchesi Thanks! Will take look early next week. Could you be so kind and provide a comparison link between the version first submitted and the updated one? Thanks!

@bstabler What's your take on the version number?

I'd suggest to stick with semantic versioning and not make it a major change if there were not breaking changes.
