### 2020-10-26 #2409 @lydialucchesi
@nuest, below is a link comparing the version that was reviewed (I pushed a few minor fixes after we submitted to JOSS and prior to your review) and the updated version of Vizumap.
https://github.com/lydialucchesi/Vizumap/compare/75628e…059c21d

Here is a link to the repository prior to the revisions.
https://github.com/lydialucchesi/Vizumap/tree/75628e28b79c52355b574e699688f53a0d9f284d
