### 2020-11-04 #2409 @GISerDaiShaoqing
@bstabler Thanks for the invitation. I'm pleased to review this paper.
