### 2020-11-26 #2409 @nuest
[I finally got to this again, I'm really sorry - I hope you'll get the second review soon to get this out.]

I took a look at [these changes](https://github.com/lydialucchesi/Vizumap/compare/75628e%E2%80%A6059c21d), compared them with [my comments](https://github.com/openjournals/joss-reviews/issues/2409#issuecomment-664007471), and subsequently updated [my review checklist](https://github.com/openjournals/joss-reviews/issues/2409#issue-647776407).

- It's great to see a CoC and contributor notes were added.
- License concerns are resolved and version numbers are aligned.
- Documentation comments were resolved/answered very thoroughly - good job! :rocket:
- Authorship concerns were answered.

Quick fixes:

- Add an empty line before the pixel map plot so that the figure is not in a line with the previous sentence ending "... (2017).".
- Add a section and one-liner to the README how to run the tests, so I can tick of the final box "Automated tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?". This is obvious for R experts, but it does not hurt to make it clear for any visitor to the repo. (If you're open to spend an hour on this, run the tests as a GitHub action and add a badge to the README, see link below.)

_With these very minor changes I'd be happy to see this submission published!_ Thank you @lydialucchesi !

Some further comments which from my perspective should not hold up further processing of the submission:

- I suggest to remove the `paper/paper.html` file, since it is not easily accessible form the repo and might be confusing. However, it's in the git history now anyway...
- `paper/makeFigures.R` should save the plots directly to the respective PNG files to make it perfectly clear which function creates with plot.
- I had a few minor warnings and notes when running `R CMD check` and running the tests (issues with CRSs), you might want to take a look at that, as you might consider submitting to CRAN at some point; might be worth adding a GitHub action and get started with CI, see https://ropenscilabs.github.io/actions_sandbox/
