+++ README.Rmd
@@ -61,11 +61,22 @@ A vignette for the `Vizumap` package is available and contains examples relating
 vignette("Vizumap")
 ```
 
+## Testing
+
+If you would like to install and run the unit tests interactively, include `INSTALL_opts = "--install-tests"` in the installation code.
+
+```
+remotes::install_github(repo = "lydialucchesi/Vizumap", build_vignettes = TRUE, force = TRUE, INSTALL_opts = "--install-tests")
+
+testthat::test_package("Vizumap", reporter = "stop")
+```
+
 ## Contribute
 
 To contribute to `Vizumap`, please follow these [guidelines](CONTRIBUTING.md).
 
 Please note that the `Vizumap` project is released with a [Contributor Code of Conduct](CONDUCT.md). By contributing to this project, you agree to abide by its terms.
+
 
 ## License
 
