### 2020-11-30 #2409 @lydialucchesi
Hi @nuest,

We have fixed the text under the pixel figure, added a README section about running tests, and included a badge to indicate whether the package is passing the R CMD check.

Thank you so much for taking the time to provide thorough reviews of Vizumap! It is greatly appreciated.

Kind regards,
Lydia
