### 2020-11-30 #2409 @GISerDaiShaoqing
The authors present an R package for visualization of uncertainty map, Vizumap. It's a very interesting and useful tool for decision-making. The package included four main types of uncertainty map. The guideline and paper are well written under the revision. Several minor revisions should be addressed as follows. @lydialucchesi

### **General checks**

- The 'Version' in DESCRIPTION file is still 1.0.0. May the authors clarify or revise it.

###  **Functionality**

- I also get a warning when running the *pixelate*. Although it seems successful, fixing it may be better?

`pix <- pixelate(ca_geo, id = "region")`

> Warning messages:
In wkt(obj) : CRS object has no comment
In wkt(obj) : CRS object has no comment
In gBuffer(geoData, byid = TRUE, width = 0) :
Spatial object is not projected; GEOS expects planar coordinates

### **Documentation**

- I think It's a good guidline for user to use the package now.
- My suggestion is providing a website guidline for this package using *pkgdown* or converting the page of *vignette* into html version.

### **Software paper**

- All the figures included in this paper lack caption. The authors should add the caption below each figure.
- The formulas should be numbered.
- In the section of 'Pixel map', the figure should be placed below the text of the paragraph.

![image](https://user-images.githubusercontent.com/26642320/100673435-bb2d0200-3363-11eb-93e4-2a303337d7bd.png)

- On page 2, the citations of *pixelate* and *biscale* seem different from other citations. Please use a uniform formatted citation style. For example, the name of the first author et al. (published year).

- Besides, each figure of the first, second, and third methods, Bivariate map, Pixel map, and Glyph map, actually consists of two sub-figures, the authors should use (a) and (b) to distinguish them.
