### 2020-12-15 242d3b0
paper edits and pkgdown site
