+++ paper/paper.md
@@ -37,7 +37,15 @@ process. Generally, this uncertainty information is shared through visualisation
 
 * The glyph map is intended to address the issue of unequal region sizes on a choropleth map. Glyphs of equal size are placed at the centroid of each region, filled with a colour that represents the estimate, and rotated depending on the degree of uncertainty.
 
-* Exceedance probability maps show the probability of exceeding a nominated threshold of concern. These probabilities can be pre-calculated and passed to the exceedance probability map function or calculated within the function given a prescribed probability distribution function, estimate, and error. For example, let $X$ denote a random variable, and $x$ is a possible value of that random variable, $X$. $F_x(x)$ is the cumulative distribution function and equals the probability that the value of $X$ is less than or equal to a specific value or threshold, $x$, so $$F_x(x) = Pr[X {\leq} x].$$ The probability that the value $X$ is greater than the specific threshold $x$ can then be written as $$Pr[X > x] = 1 - F_x(x) = 1 - Pr[X {\leq} x].$$ In R, assuming an exponential distribution, this can be evaluated through the following expressions: `1 - pexp(q, rate)` or `pexp(q, rate, lower.tail = FALSE)`.
+* Exceedance probability maps show the probability of exceeding a nominated threshold of concern. These probabilities can be pre-calculated and passed to the exceedance probability map function or calculated within the function given a prescribed probability distribution function, estimate, and error. For example, let $X$ denote a random variable, and $x$ is a possible value of that random variable, $X$. $F_x(x)$ is the cumulative distribution function and equals the probability that the value of $X$ is less than or equal to a specific value or threshold, $x$, so 
+\begin{equation}\label{eq:eq1}
+{F_x(x) = Pr[X {\leq} x].}
+\end{equation}
+The probability that the value $X$ is greater than the specific threshold $x$ can then be written as 
+\begin{equation}\label{eq:eq2}
+{Pr[X > x] = 1 - F_x(x) = 1 - Pr[X {\leq} x].}
+\end{equation}
+In R, assuming an exponential distribution, this can be evaluated through the following expressions: `1 - pexp(q, rate)` or `pexp(q, rate, lower.tail = FALSE)`.
 
 Other related R packages include `pixelate` and `biscale`. The `pixelate` package explores the use of pixelation for uncertainty visualisation on isopleth maps [@pixelatePackage], and the `biscale` package can be used to generate bivariate maps for two variables [@biscalePackage].
 
@@ -47,27 +55,27 @@ As an illustration below, we use `Vizumap` to visualise estimated pollutant load
 
 ## Bivariate map
 
-The bivariate map of the upper Burdekin catchment in Queensland, Australia, depicts predicted sediment concentrations and prediction uncertainty. In this example, a custom colour palette is created and then used to build the bivariate map and colour key. The bivariate bins are defined using terciles.
+Figure \ref{bivMap} is a bivariate map of the upper Burdekin catchment in Queensland, Australia. It visualises predicted sediment concentrations and prediction uncertainty. In this example, a custom colour palette is created and then used to build the bivariate map and colour key. The bivariate bins are defined using terciles.
 
-![](bivariateMap.png){width=4in height=4in}
+![Bivariate map and bivariate key.\label{bivMap}](bivariateMap.png){width=4in height=4in}
 
 ## Pixel map
 
[…]
