+++ pkgdown.yml
@@ -1,0 +1,7 @@
+pandoc: 2.7.3
+pkgdown: 1.6.1
+pkgdown_sha: ~
+articles:
+  Vizumap: Vizumap.html
+last_built: 2020-12-15T05:43Z
+
