### 2020-12-15 cd86268
pkgdown formatting
