+++ _pkgdown.yml
@@ -1,0 +1,47 @@
+pandoc: '2.5'
+pkgdown: 1.3.0
+reference:
+- title: Data and maps
+  contents:
+  - '`us_data`'
+  - '`us_geo`'
+  - '`UB_shp`'
+  - '`UB_tss`'
+- title: Formatting functions
+  desc: Functions for preparing data frames and maps for the building functions
+  contents:
+  - '`read.uv`'
+  - '`pixelate`'
+- title: Building functions
+  desc: Functions for making maps, keys, colour palettes, and animations
+  contents:
+  - '`build_bmap`'
+  - '`build_pmap`'
+  - '`build_gmap`'
+  - '`build_emap`'
+  - '`build_bkey`'
+  - '`build_gkey`'
+  - '`build_palette`'
+  - '`animate`'
+- title: Viewing functions
+  desc: Functions for plotting objects made with the building functions
+  contents:
+  - '`view`'
+  - '`attach_key`'
+- title: Helper functions
+  contents:
+  - '`findNbounds`'
+  - '`pdist`'
+
+
+navbar:
+  type: default
+  left:
+  - text: Reference
+    href: reference/index.html
+  - text: Vignettes
+    menu:
+    - text: Using Vizumap
+      href: articles/Vizumap.html
+  - text: News
+    href: news/index.html
