+++ README.Rmd
@@ -19,6 +19,8 @@ knitr::opts_chunk$set(
 
 [![R build status](https://github.com/lydialucchesi/Vizumap/workflows/R-CMD-check/badge.svg)](https://github.com/lydialucchesi/Vizumap/actions)
 
+`Vizumap` pkgdown site with vignette: https://lydialucchesi.github.io/Vizumap/.
+
 ## Installation
 
 You can install a development version of the `Vizumap` package using the command below.
@@ -54,7 +56,7 @@ The final map-based exploration is through exceedance probabilities, which are v
 
 ## Examples
 
-A vignette for the `Vizumap` package is available and contains examples relating to each of the visualisation methods.
+A for the `Vizumap` package is available and contains examples relating to each of the visualisation methods.
 
 ```
 vignette("Vizumap")
