+++ _pkgdown.yml
@@ -1,20 +1,19 @@
-pandoc: '2.5'
-pkgdown: 1.3.0
 reference:
 - title: Data and maps
-  contents:
+  desc:
+- contents:
   - '`us_data`'
   - '`us_geo`'
   - '`UB_shp`'
   - '`UB_tss`'
 - title: Formatting functions
   desc: Functions for preparing data frames and maps for the building functions
-  contents:
+- contents:
   - '`read.uv`'
   - '`pixelate`'
 - title: Building functions
   desc: Functions for making maps, keys, colour palettes, and animations
-  contents:
+- contents:
   - '`build_bmap`'
   - '`build_pmap`'
   - '`build_gmap`'
@@ -25,23 +24,20 @@ reference:
   - '`animate`'
 - title: Viewing functions
   desc: Functions for plotting objects made with the building functions
-  contents:
+- contents:
   - '`view`'
   - '`attach_key`'
 - title: Helper functions
-  contents:
+  desc:
+- contents:
   - '`findNbounds`'
   - '`pdist`'
 
-
 navbar:
-  type: default
   left:
+  - text: Vignette
+    href: articles/Vizumap.html
   - text: Reference
     href: reference/index.html
-  - text: Vignettes
-    menu:
-    - text: Using Vizumap
-      href: articles/Vizumap.html
   - text: News
     href: news/index.html
