### 2020-12-20 b91007f
updates to website, paper, and pixelate function
