+++ paper/paper.md
@@ -33,7 +33,7 @@ process. Generally, this uncertainty information is shared through visualisation
 
 * For the bivariate choropleth map, two colour scales are merged to create a bivariate grid that can encode the estimates and errors at once, making it possible to fill each region with a single colour that represents both values.
 
-* The pixel map is generated by dividing each region into small pixels, which are then filled with a colour value sampled from the colour gradient range that is encompassed by the estimate plus and minus the error. Areas of greater uncertainty appear pixelated while areas of less uncertainty appear smooth, as though filled by only a single colour. These pixel maps can be animated so that areas of high uncertainty flicker while areas of low uncertainty appear static to the eye.
+* The pixel map is generated by dividing each region into small pixels, which are then filled with a colour value sampled from the colour gradient range that is encompassed by the estimate plus or minus the error. Areas of greater uncertainty appear pixelated while areas of less uncertainty appear smooth, as though filled by only a single colour. These pixel maps can be animated so that areas of high uncertainty flicker while areas of low uncertainty appear static to the eye.
 
 * The glyph map is intended to address the issue of unequal region sizes on a choropleth map. Glyphs of equal size are placed at the centroid of each region, filled with a colour that represents the estimate, and rotated depending on the degree of uncertainty.
 
@@ -55,27 +55,27 @@ As an illustration below, we use `Vizumap` to visualise estimated pollutant load
 
 ## Bivariate map
 
-Figure \ref{bivMap} is a bivariate map of the upper Burdekin catchment in Queensland, Australia. It visualises predicted sediment concentrations and prediction uncertainty. In this example, a custom colour palette is created and then used to build the bivariate map and colour key. The bivariate bins are defined using terciles.
+Figure \ref{bivMap} is a bivariate map of the upper Burdekin catchment in Queensland, Australia. It visualises predicted total suspended sediment (TSS) concentrations and prediction uncertainty. In this example, a custom colour palette is created and then used to build the bivariate map and colour key. The bivariate bins are defined using terciles.
 
-![Bivariate map and bivariate key.\label{bivMap}](bivariateMap.png){width=4in height=4in}
+![Bivariate map and bivariate key for the upper Burdekin catchment in Queensland, Australia, showing the total suspended sediment (TSS) concentration in mg/L.\label{bivMap}](bivariateMap.png){width=4in height=4in}
 
 ## Pixel map
 
-Figure \ref{pixMap} visualises the uncertainty of TSS predictions, while giving a general idea of estimated TSS loads. For a closer look at the pixelation, a subset of regions is included to the right of the map (sub-figure B). The colours filling the pixels in each region were sampled from the estimate's relative frequency distribution. This pixel map can be animated so that the pixels flicker between sampled values. If the map below were to be animated, the areas that appear most pixelated in the static map would, correspondingly, have the most visible movement among pixels in the animated map. Movement among pixels in areas of low uncertainty would be hard to detect due to the minimal differences between the sampled values of orange. An example of an animated pixel map can be found in @vizMethod.
[…]
