+++ R/pixelate.R
@@ -3,7 +3,7 @@
 #'This function divides a shapefile into pixels so it can be used to create a
 #'pixel map with \code{\link{build_pmap}}.
 #'
-#'This function can take several minutes to run depending on the size of the shapefile.
+#'This function can take several minutes to run depending on the size of the shapefile. Within this function, the projection of the spatial object is removed and then returned to the original projection.
 #'
 #'@param geoData An object of class "SpatialPolygons" or
 #'  "SpatialPolygonsDataFrame".
@@ -31,92 +31,103 @@
 
 
 
-pixelate <- function(geoData = NULL, file = NULL, layer = NULL, pixelSize = 2, id = "id") {
+pixelate <-
+  function(geoData = NULL,
+           file = NULL,
+           layer = NULL,
+           pixelSize = 2,
+           id = "id") {
+    # check for geoData or file
+    if (is.null(geoData) & is.null(file))
+      stop("One of geoData or file needs to be supplied to this function.\n")
 
+    # check class of geoData
+    if (is.null(file) &
+        (!(
+          class(geoData) %in% c("SpatialPolygons", "SpatialPolygonsDataFrame")
+        )))
+      stop("geoData must be one of class SpatialPolygons or SpatialPolygonsDataFrame.\n")
 
+    # check that layer is entered with file
+    if (!is.null(file) & is.null(layer))
+      stop("Layer needs to be supplied if file is supplied.\n")
 
-  # check for geoData or file
-  if (is.null(geoData) & is.null(file))
-    stop("One of geoData or file needs to be supplied to this function.\n")
-
-  # check class of geoData
-  if (is.null(file) & (!(class(geoData) %in% c("SpatialPolygons", "SpatialPolygonsDataFrame"))))
-    stop("geoData must be one of class SpatialPolygons or SpatialPolygonsDataFrame.\n")
-
-  # check that layer is entered with file
-  if (!is.null(file) & is.null(layer))
-    stop("Layer needs to be supplied if file is supplied.\n")
-
-  # find projection of geo object
-  if (!is.null(geoData))
-    x <- proj4string(geoData)
-  else {
-    geoData <- readOGR(file, layer = layer)
-    x <- proj4string(geoData)
-  }
-
-  if (!is.na(x)) {
-    if (!is.null(file)) {
-      geoData <- SpatialPolygons(geoData@polygons, proj4string = geoData@proj4string)
+    # find projection of geo object
+    if (!is.null(geoData))
+      x <- proj4string(geoData)
+    else {
+      geoData <- readOGR(file, layer = layer)
+      x <- proj4string(geoData)
     }
-    if (class(geoData) == "SpatialPolygonsDataFrame") {
-      geoData <- SpatialPolygons(geoData@polygons, proj4string = geoData@proj4string)
+
+    if (!is.na(x)) {
+      if (!is.null(file)) {
+        geoData <-
+          SpatialPolygons(geoData@polygons, proj4string = geoData@proj4string)
+      }
+      if (class(geoData) == "SpatialPolygonsDataFrame") {
+        geoData <-
+          SpatialPolygons(geoData@polygons, proj4string = geoData@proj4string)
+      }
     }
+
+    # remove projection
+    pol <- "POLYGON((0 0, 0 1, 1 1, 1 0, 0 0))"
+    poly <- readWKT(pol)
+    geoDataProjNA <- geoData
+    proj4string(geoDataProjNA) <- proj4string(poly)
+
+    # create definitive space
[…]
