+++ README.Rmd
@@ -90,7 +90,7 @@ How do you include uncertainty on a map displaying areal data estimates?
 
 Over the course of a year, they put together three methods for visualising uncertainty in spatial statistics: the bivariate choropleth map, the pixel map, and the glyph map. By mid-2017, there were maps, and there was a lot of R code, but there was not a tool that others could use to easily make these types of maps, too. That’s when statistician Petra Kuhnert recommended developing an R package. Over the course of a month, Petra and Lydia developed Vizumap (originally named VizU) at CSIRO Data61 in Canberra, Australia. Since then, the package has been expanded to include exceedance probability maps, an uncertainty visualisation method developed by Petra while working on a Great Barrier Reef (GBR) project.
 
-Vizumap has been used to visualise the uncertainty of American Community Survey estimates, the prediction errors of sediment load estimates in a GBR catchment, and most recently the [uncertainty of estimated locust densities in Australia](https://www.nature.com/articles/s41598-020-73897-1/figures/4). We would like to assemble a Vizumap gallery that showcases different applications of the package’s mapping methods. If you use Vizumap to visualise uncertainty, please feel free to send the map our way. We would like to see it!
+Vizumap has been used to visualise the uncertainty of American Community Survey estimates, the prediction errors of sediment estimates in a GBR catchment, and most recently the [uncertainty of estimated locust densities in Australia](https://www.nature.com/articles/s41598-020-73897-1/figures/4). We would like to assemble a Vizumap gallery that showcases different applications of the package’s mapping methods. If you use Vizumap to visualise uncertainty, please feel free to send the map our way. We would like to see it!
 
 ## References
 

[…]
