+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 1.6.1
 pkgdown_sha: ~
 articles:
   Vizumap: Vizumap.html
-last_built: 2020-12-15T11:46Z
+last_built: 2020-12-20T23:38Z
 
