+++ paper/paper.md
@@ -33,7 +33,7 @@ process. Generally, this uncertainty information is shared through visualisation
 
 * For the bivariate choropleth map, two colour scales are merged to create a bivariate grid that can encode the estimates and errors at once, making it possible to fill each region with a single colour that represents both values.
 
-* The pixel map is generated by dividing each region into small pixels, which are then filled with a colour value sampled from the colour gradient range that is encompassed by the estimate plus or minus the error. Areas of greater uncertainty appear pixelated while areas of less uncertainty appear smooth, as though filled by only a single colour. These pixel maps can be animated so that areas of high uncertainty flicker while areas of low uncertainty appear static to the eye.
+* The pixel map is generated by dividing each region into small pixels, which are then filled with a colour value sampled from the colour gradient range. The method for assigning colours to pixels depends on the type of distribution specified. Areas of greater uncertainty appear pixelated while areas of less uncertainty appear smooth, as though filled by only a single colour. These pixel maps can be animated so that areas of high uncertainty flicker while areas of low uncertainty appear static to the eye.
 
 * The glyph map is intended to address the issue of unequal region sizes on a choropleth map. Glyphs of equal size are placed at the centroid of each region, filled with a colour that represents the estimate, and rotated depending on the degree of uncertainty.
 
@@ -67,7 +67,7 @@ Figure \ref{pixMap} visualises the uncertainty of TSS predictions, while giving 
 
 ## Glyph map
 
-Figure \ref{glyphMap} depicts estimated TSS concetrations and the uncertainty of these predictions. The colour filling each glyph (located at the region centroid) represents the estimate, and the rotation of the glyph represents the uncertainty.
+Figure \ref{glyphMap} depicts estimated TSS concentrations and the uncertainty of these predictions. The colour filling each glyph (located at the region centroid) represents the estimate, and the rotation of the glyph represents the uncertainty.
 
 ![Glyph map and glyph key showing the TSS predictions and uncertainties for the upper Burdekin catchment in Queensland, Australia.\label{glyphMap}](glyphMap.png){width=4in height=4in}
 
