### 2020-12-21 #2409 @lydialucchesi
Dear @GISerDaiShaoqing ,

Thank you for taking the time to review Vizumap. We have addressed the suggested revisions, and below are comments regarding each revision. Additionally, all edits to the package and paper have been documented here: https://github.com/lydialucchesi/Vizumap/blob/master/paper/revisions.md.

**General checks**

We accidentally submitted the wrong version number in the JOSS submission form. The correct version was 1.0.0. However, given all of the recent changes, as of today we have updated the version to 1.1.0. The description file has been updated accordingly to 1.1.0.

**Functionality**

To deal with the pixelate warnings, we now remove the projection in order to “gBuffer” and “geoaxe::chop” and then return the spatial object to its original projection. A comment regarding this has been added to the pixelate function documentation so users are aware.

**Documentation**

A pkgdown site has been generated, and a link to the site (https://lydialucchesi.github.io/Vizumap/) is provided on the README page.

We have also added a new section to the README called “History of Vizumap.” We felt this would be a good way to document how the collaboration came about and the different author contributions requested earlier in the JOSS review process.

**Software paper**

Captions have been added for each figure, and formulas have been numbered. The pixel map is now below the pixel paragraph.

When including in-text citations, we followed the JOSS guidelines provided here: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography. The in-text citation format we chose depends on the sentence structure, and both formats are supported by JOSS. Please let us know if you would like us to reword the sentences.

We have updated the pixel map to include sub-figure labels. If possible, we would prefer to not add sub-figure labels to the bivariate and glyph maps as we would like the map and map key to be thought of as just one figure and not two separate figures.

We greatly appreciate your thoughtful review of Vizumap and are excited about the changes/additions, especially the new pkgdown site! Please let us know if there is any additional information we can provide regarding the revisions for this second JOSS review.

Kind regards,
Lydia
