### 2020-12-21 #2409 @lydialucchesi
Following is a link comparing the repo that was reviewed and the updated version: https://github.com/lydialucchesi/Vizumap/compare/8b28d0f...1b96771.
