+++ DESCRIPTION
@@ -31,6 +31,7 @@ Imports:
     sp,
     spbabel,
     maps
+Remotes: url::https://github.com/ropensci-archive/geoaxe/archive/master.zip
 License: GPL-3
 LazyLoad: TRUE
 LazyData: TRUE
