### 2021-02-01 #2409 @lydialucchesi
Thank you for reviewing Vizumap, @GISerDaiShaoqing.

We reviewed the proofs and made a few small edits.

- Capitalised all instances of "figure"
- Changed Sub-figure B to Figure 2(B) (likewise for Figure 2(A))
- Added "for the second author" in acknowledgements
- Changed "uncertainty" to "uncertainties" in two places

@bstabler, what are the next steps in the review process?

King regards,
Lydia
