+++ README.md
@@ -88,7 +88,7 @@ to abide by its terms.
 
 ## License
 
-`Vizumap` version 1.1.0 is licensed under [GPLv3](LICENSE.md).
+`Vizumap` version 1.2.0 is licensed under [GPLv3](LICENSE.md).
 
 ## History of Vizumap
 
