### 2021-02-22 #2409 @lydialucchesi
Dear @arfon,

Thank you for your message. We have created a new release of Vizumap (v1.2.0) that includes the changes resulting from the JOSS review. We have also made an archive of the software on Zenodo. We first archived the R package on Zenodo in 2017. The latest archive for JOSS is linked to the original archive. Will that be okay?

A link to the Zenodo archive of Vizumap:
https://doi.org/10.5281/zenodo.4554558

The DOI:
10.5281/zenodo.4554558

Also, is it possible to get a copy of the most recent paper proof? We have made several small changes since the last proof was generated.

Kind regards,
Lydia
