### 2021-02-22 #2409 @arfon
@lydialucchesi - this is the [latest proof](https://github.com/openjournals/joss-papers/blob/b271ac47cacfb57db85d04dd080332f41744ab96/joss.02409/10.21105.joss.02409.pdf) based on all the changes to date.
