### 2021-03-05 #2409 @danielskatz
As far as I can tell, this is ready to publish but got dropped by us (JOSS) somehow, so I'm going to go ahead.
